requirements.txt
Django==4.2.8
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-decouple==3.8
Pillow==10.1.0
reportlab==4.0.7
requests==2.31.0
django-filter==23.4
whitenoise==6.6.0
gunicorn==21.2.0
fdb==2.0.4
cryptography==46.0.5
openpyxl==3.1.2
